# swagger.yaml
basePath: /api/v1
consumes:
- application/json
definitions:
  Account:
    properties:
      created_at:
        format: date-time
        type: string
        x-go-name: CreatedAt
      deleted_at:
        $ref: '#/definitions/DeletedAt'
      id:
        $ref: '#/definitions/UUID'
      name:
        type: string
        x-go-name: Name
      tenants:
        items:
          $ref: '#/definitions/Tenant'
        type: array
        x-go-name: Tenants
      updated_at:
        format: date-time
        type: string
        x-go-name: UpdatedAt
      users:
        items:
          $ref: '#/definitions/User'
        type: array
        x-go-name: Users
    type: object
    x-go-package: github.com/solabsafrica/afrikanest/model
...
  CreateLeaseChargePaymentRequestBody:
    properties:
      amount:
        description: 'in: body'
        format: int64
        type: integer
        x-go-name: Amount
      paymentDate:
        type: string
        x-go-name: PaymentDate
      sentEmailToTenant:
        type: boolean
        x-go-name: SentEmailToTenant
    required:
    - amount
    type: object
    x-go-package: github.com/solabsafrica/afrikanest/restful/request
  CreateLeaseChargeRequestBody:
    properties:
      amount:
        description: 'in: body'
        format: int64
        type: integer
        x-go-name: Amount
      charge_type:
        $ref: '#/definitions/LeaseChargeTypeEmun'
      description:
        description: 'in: body'
        type: string
        x-go-name: Description
      due_date:
        description: 'in: body'
        type: string
        x-go-name: DueDate
      notify_tenant:
        description: 'in: body'
        type: boolean
        x-go-name: NotifyTenant
    required:
    - amount
    - description
    - due_date
    - notify_tenant
    - charge_type
    type: object
    x-go-package: github.com/solabsafrica/afrikanest/restful/request
  CreateLeaseRequest:
    properties:
      invite:
        type: boolean
        x-go-name: InviteTenants
      rent:
        format: int64
        type: integer
        x-go-name: MonthlyRent
      rent_due_day_of_month:
        format: int64
        type: integer
        x-go-name: RentDueDayOfMonth
      security_deposit:
        format: int64
        type: integer
        x-go-name: SecurityDeposit
      start_date:
        type: string
        x-go-name: StartDate
      tenants:
        items:
          $ref: '#/definitions/CreateTenantRequest'
        type: array
        x-go-name: Tenants
      unit_id:
        type: string
        x-go-name: UnitID
    type: object
    x-go-package: github.com/solabsafrica/afrikanest/restful/request
  CreateLeaseResponse:
    properties:
      created_at:
        format: date-time
        type: string
        x-go-name: CreatedAt
      id:
        type: string
        x-go-name: ID
    type: object
    x-go-package: github.com/solabsafrica/afrikanest/restful/response
...
  CreateTenantRequest:
    properties:
      email:
        type: string
        x-go-name: Email
      first_name:
        type: string
        x-go-name: FirstName
      last_name:
        type: string
        x-go-name: LastName
      phone_number:
        type: string
        x-go-name: PhoneNumber
      send_invite:
        type: boolean
        x-go-name: SendInvite
    type: object
    x-go-package: github.com/solabsafrica/afrikanest/restful/request
...
  DeletedAt:
    $ref: '#/definitions/NullTime'
...
  EncryptedPassword:
    type: string
    x-go-package: github.com/solabsafrica/afrikanest/model
  ErrorResponse:
    properties:
      message:
        type: string
        x-go-name: Message
      stack_trace:
        type: string
        x-go-name: StackTrace
      title:
        type: string
        x-go-name: Title
      type:
        type: string
        x-go-name: Type
    type: object
    x-go-package: github.com/solabsafrica/afrikanest/restful/response
...
  Lease:
    properties:
      created_at:
        format: date-time
        type: string
        x-go-name: CreatedAt
      deleted_at:
        $ref: '#/definitions/DeletedAt'
      end_date:
        format: date-time
        type: string
        x-go-name: EndDate
      id:
        $ref: '#/definitions/UUID'
      lease_charges:
        items:
          $ref: '#/definitions/LeaseCharge'
        type: array
        x-go-name: LeaseCharge
      lease_number:
        type: string
        x-go-name: LeaseNumber
      lease_status:
        $ref: '#/definitions/LeaseStatusType'
      rent:
        format: int64
        type: integer
        x-go-name: MonthlyRent
      rent_due_day_of_month:
        format: int64
        type: integer
        x-go-name: RentDueDayOfMonth
      security_deposit:
        format: int64
        type: integer
        x-go-name: SecurityDeposit
      start_date:
        format: date-time
        type: string
        x-go-name: StartDate
      tenants:
        items:
          $ref: '#/definitions/Tenant'
        type: array
        x-go-name: Tenants
      unit:
        $ref: '#/definitions/Unit'
      unit_id:
        $ref: '#/definitions/UUID'
      updated_at:
        format: date-time
        type: string
        x-go-name: UpdatedAt
    type: object
    x-go-package: github.com/solabsafrica/afrikanest/model
  LeaseBalanceResponse:
    properties:
      balance:
        format: int64
        type: integer
        x-go-name: Balance
    type: object
    x-go-package: github.com/solabsafrica/afrikanest/restful/response
  LeaseCharge:
    properties:
      amount:
        format: int64
        type: integer
        x-go-name: Amount
      charge_type:
        $ref: '#/definitions/LeaseChargeTypeEmun'
      created_at:
        format: date-time
        type: string
        x-go-name: CreatedAt
      deleted_at:
        $ref: '#/definitions/DeletedAt'
      description:
        type: string
        x-go-name: Description
      due_date:
        format: date-time
        type: string
        x-go-name: DueDate
      id:
        $ref: '#/definitions/UUID'
      lease:
        $ref: '#/definitions/Lease'
      lease_charge_type_id:
        $ref: '#/definitions/UUID'
      lease_charges_payment:
        items:
          $ref: '#/definitions/LeaseChargePayment'
        type: array
        x-go-name: LeaseChargesPayments
      lease_id:
        $ref: '#/definitions/UUID'
      name:
        type: string
        x-go-name: Name
      recived_amount:
        format: int64
        type: integer
        x-go-name: RecivedAmount
      remaining_amount:
        format: int64
        type: integer
        x-go-name: RemainingAmount
      updated_at:
        format: date-time
        type: string
        x-go-name: UpdatedAt
    type: object
    x-go-package: github.com/solabsafrica/afrikanest/model
  LeaseChargePayment:
    properties:
      amount:
        format: int64
        type: integer
        x-go-name: Amount
      created_at:
        format: date-time
        type: string
        x-go-name: CreatedAt
      deleted_at:
        $ref: '#/definitions/DeletedAt'
      id:
        $ref: '#/definitions/UUID'
      lease_charge:
        $ref: '#/definitions/LeaseCharge'
      lease_charge_id:
        $ref: '#/definitions/UUID'
      payment_date:
        format: date-time
        type: string
        x-go-name: PaymentDate
      updated_at:
        format: date-time
        type: string
        x-go-name: UpdatedAt
    type: object
    x-go-package: github.com/solabsafrica/afrikanest/model
  LeaseChargePaymentResponse:
    properties:
      amount:
        format: int64
        type: integer
        x-go-name: Amount
      created_at:
        format: date-time
        type: string
        x-go-name: CreatedAt
      deleted_at:
        $ref: '#/definitions/DeletedAt'
      id:
        $ref: '#/definitions/UUID'
      lease_charge:
        $ref: '#/definitions/LeaseCharge'
      lease_charge_id:
        $ref: '#/definitions/UUID'
      payment_date:
        format: date-time
        type: string
        x-go-name: PaymentDate
      updated_at:
        format: date-time
        type: string
        x-go-name: UpdatedAt
    type: object
    x-go-package: github.com/solabsafrica/afrikanest/restful/response
  LeaseChargeResponse:
    properties:
      amount:
        format: int64
        type: integer
        x-go-name: Amount
      charge_type:
        $ref: '#/definitions/LeaseChargeTypeEmun'
      created_at:
        format: date-time
        type: string
        x-go-name: CreatedAt
      deleted_at:
        $ref: '#/definitions/DeletedAt'
      description:
        type: string
        x-go-name: Description
      due_date:
        format: date-time
        type: string
        x-go-name: DueDate
      id:
        $ref: '#/definitions/UUID'
      lease:
        $ref: '#/definitions/Lease'
      lease_charge_type_id:
        $ref: '#/definitions/UUID'
      lease_charges_payment:
        items:
          $ref: '#/definitions/LeaseChargePayment'
        type: array
        x-go-name: LeaseChargesPayments
      lease_id:
        $ref: '#/definitions/UUID'
      name:
        type: string
        x-go-name: Name
      recived_amount:
        format: int64
        type: integer
        x-go-name: RecivedAmount
      remaining_amount:
        format: int64
        type: integer
        x-go-name: RemainingAmount
      updated_at:
        format: date-time
        type: string
        x-go-name: UpdatedAt
    type: object
    x-go-package: github.com/solabsafrica/afrikanest/restful/response
  LeaseChargeTypeEmun:
    type: string
    x-go-package: github.com/solabsafrica/afrikanest/model
  LeaseResponse:
    properties:
      created_at:
        format: date-time
        type: string
        x-go-name: CreatedAt
      deleted_at:
        $ref: '#/definitions/DeletedAt'
      end_date:
        format: date-time
        type: string
        x-go-name: EndDate
      id:
        $ref: '#/definitions/UUID'
      lease_charges:
        items:
          $ref: '#/definitions/LeaseCharge'
        type: array
        x-go-name: LeaseCharge
      lease_number:
        type: string
        x-go-name: LeaseNumber
      lease_status:
        $ref: '#/definitions/LeaseStatusType'
      rent:
        format: int64
        type: integer
        x-go-name: MonthlyRent
      rent_due_day_of_month:
        format: int64
        type: integer
        x-go-name: RentDueDayOfMonth
      security_deposit:
        format: int64
        type: integer
        x-go-name: SecurityDeposit
      start_date:
        format: date-time
        type: string
        x-go-name: StartDate
      tenants:
        items:
          $ref: '#/definitions/Tenant'
        type: array
        x-go-name: Tenants
      unit:
        $ref: '#/definitions/Unit'
      unit_id:
        $ref: '#/definitions/UUID'
      updated_at:
        format: date-time
        type: string
        x-go-name: UpdatedAt
    type: object
    x-go-package: github.com/solabsafrica/afrikanest/restful/response
  LeaseStatusType:
    type: string
    x-go-package: github.com/solabsafrica/afrikanest/model
...
  NullTime:
    description: |-
      NullTime implements the Scanner interface so
      it can be used as a scan destination, similar to NullString.
    properties:
      Time:
        format: date-time
        type: string
      Valid:
        type: boolean
    title: NullTime represents a time.Time that may be null.
    type: object
    x-go-package: database/sql
  Pagination:
    properties:
      page:
        format: int64
        type: integer
        x-go-name: Page
      page_size:
        format: int64
        type: integer
        x-go-name: PageSize
      total:
        format: int64
        type: integer
        x-go-name: Total
    type: object
    x-go-package: github.com/solabsafrica/afrikanest/restful/response
  Property:
    properties:
      created_at:
        format: date-time
        type: string
        x-go-name: CreatedAt
      deleted_at:
        $ref: '#/definitions/DeletedAt'
      description:
        type: string
        x-go-name: Description
      id:
        $ref: '#/definitions/UUID'
      name:
        type: string
        x-go-name: Name
      owner_id:
        $ref: '#/definitions/UUID'
      units:
        items:
          $ref: '#/definitions/Unit'
        type: array
        x-go-name: Units
      updated_at:
        format: date-time
        type: string
        x-go-name: UpdatedAt
    type: object
    x-go-package: github.com/solabsafrica/afrikanest/model
...
  Tenant:
    properties:
      account:
        $ref: '#/definitions/Account'
      account_id:
        $ref: '#/definitions/UUID'
      created_at:
        format: date-time
        type: string
        x-go-name: CreatedAt
      created_by:
        $ref: '#/definitions/User'
      created_by_id:
        $ref: '#/definitions/UUID'
      deleted_at:
        $ref: '#/definitions/DeletedAt'
      email:
        type: string
        x-go-name: Email
      id:
        $ref: '#/definitions/UUID'
      invite_accepted:
        format: date-time
        type: string
        x-go-name: InviteAccepted
      invite_sent:
        format: date-time
        type: string
        x-go-name: InviteSent
      invite_token:
        type: string
        x-go-name: InviteToken
      last_name:
        type: string
        x-go-name: LastName
      lease:
        $ref: '#/definitions/Lease'
      lease_id:
        $ref: '#/definitions/UUID'
      name:
        type: string
        x-go-name: FirstName
      phone:
        type: string
        x-go-name: PhoneNumber
      updated_at:
        format: date-time
        type: string
        x-go-name: UpdatedAt
      user:
        $ref: '#/definitions/User'
      user_id:
        $ref: '#/definitions/UUID'
    type: object
    x-go-package: github.com/solabsafrica/afrikanest/model
  TenantsResponse:
    properties:
      pagination:
        $ref: '#/definitions/Pagination'
      tenants:
        items:
          $ref: '#/definitions/Tenant'
        type: array
        x-go-name: Tenants
    type: object
    x-go-package: github.com/solabsafrica/afrikanest/restful/response
  UUID:
    description: |-
      A UUID is a 128 bit (16 byte) Universal Unique IDentifier as defined in RFC
      4122.
    items:
      format: uint8
      type: integer
    type: array
    x-go-package: github.com/google/uuid
  Unit:
    properties:
      created_at:
        format: date-time
        type: string
        x-go-name: CreatedAt
      default_rent:
        format: int64
        type: integer
        x-go-name: DefaultRent
      deleted_at:
        $ref: '#/definitions/DeletedAt'
      description:
        type: string
        x-go-name: Description
      id:
        $ref: '#/definitions/UUID'
      leases:
        items:
          $ref: '#/definitions/Lease'
        type: array
        x-go-name: Leases
      name:
        type: string
        x-go-name: Name
      property:
        $ref: '#/definitions/Property'
      property_id:
        $ref: '#/definitions/UUID'
      updated_at:
        format: date-time
        type: string
        x-go-name: UpdatedAt
    type: object
    x-go-package: github.com/solabsafrica/afrikanest/model
...
  User:
    properties:
      Accounts:
        items:
          $ref: '#/definitions/Account'
        type: array
      Email:
        type: string
      EncryptedPassword:
        $ref: '#/definitions/EncryptedPassword'
      FirstName:
        type: string
      LastName:
        type: string
      Phone:
        type: string
      created_at:
        format: date-time
        type: string
        x-go-name: CreatedAt
      deleted_at:
        $ref: '#/definitions/DeletedAt'
      id:
        $ref: '#/definitions/UUID'
      updated_at:
        format: date-time
        type: string
        x-go-name: UpdatedAt
    type: object
    x-go-package: github.com/solabsafrica/afrikanest/model
host: localhost:8081
info:
  description: Documentation of AFRIKANEST API.
  title: afrikanest.
  version: 1.0.0
paths:
  /leases:
    post:
      consumes:
      - application/json
      description: Create a lease
      operationId: createLease
      parameters:
      - in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/CreateLeaseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: CreateLeaseResponse
          schema:
            $ref: '#/definitions/CreateLeaseResponse'
        "401":
          description: ErrorResponse
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: ErrorResponse
          schema:
            $ref: '#/definitions/ErrorResponse'
      tags:
      - lease
  /leases/{id}:
    get:
      consumes:
      - application/json
      description: Get a lease
      operationId: getLease
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      produces:
      - application/json
      responses:
        "201":
          description: LeaseResponse
          schema:
            $ref: '#/definitions/LeaseResponse'
        "401":
          description: ErrorResponse
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: ErrorResponse
          schema:
            $ref: '#/definitions/ErrorResponse'
      tags:
      - lease
  /leases/{id}/balance:
    get:
      consumes:
      - application/json
      description: Get a lease balance
      operationId: getLeaseBalance
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      produces:
      - application/json
      responses:
        "200":
          description: LeaseBalanceResponse
          schema:
            $ref: '#/definitions/LeaseBalanceResponse'
        "401":
          description: ErrorResponse
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: ErrorResponse
          schema:
            $ref: '#/definitions/ErrorResponse'
      tags:
      - lease
  /leases/{id}/charges:
    post:
      consumes:
      - application/json
      description: Create a lease charge
      operationId: createLeaseCharge
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      - in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/CreateLeaseChargeRequestBody'
      produces:
      - application/json
      responses:
        "201":
          description: LeaseChargeResponse
          schema:
            $ref: '#/definitions/LeaseChargeResponse'
        "401":
          description: ErrorResponse
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: ErrorResponse
          schema:
            $ref: '#/definitions/ErrorResponse'
      tags:
      - lease
  /leases/{id}/charges/{leaseChargeId}:
    get:
      consumes:
      - application/json
      description: Get a lease charges
      operationId: getLeaseCharge
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      - in: path
        name: leaseChargeId
        required: true
        type: string
        x-go-name: LeaseChargeID
      produces:
      - application/json
      responses:
        "200":
          description: LeaseChargeResponse
          schema:
            $ref: '#/definitions/LeaseChargeResponse'
        "401":
          description: ErrorResponse
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: ErrorResponse
          schema:
            $ref: '#/definitions/ErrorResponse'
      tags:
      - lease
  /leases/{id}/charges/{leaseChargeId}/payments:
    post:
      consumes:
      - application/json
      description: Create a lease charge
      operationId: createLeaseChargePayment
      parameters:
      - in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/CreateLeaseChargePaymentRequestBody'
      - in: path
        name: leaseChargeId
        required: true
        type: string
        x-go-name: LeaseChargeID
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      produces:
      - application/json
      responses:
        "201":
          description: LeaseChargePaymentResponse
          schema:
            $ref: '#/definitions/LeaseChargePaymentResponse'
        "401":
          description: ErrorResponse
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: ErrorResponse
          schema:
            $ref: '#/definitions/ErrorResponse'
      tags:
      - lease
...
  /tenants:
    get:
      consumes:
      - application/json
      description: Get all tenants
      operationId: getTenants
      produces:
      - application/json
      responses:
        "200":
          description: TenantsResponse
          schema:
            $ref: '#/definitions/TenantsResponse'
        "401":
          description: ErrorResponse
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: ErrorResponse
          schema:
            $ref: '#/definitions/ErrorResponse'
      schemes:
      - http
      - https
      tags:
      - Tenants
...
produces:
- application/json
schemes:
- http
security:
- bearerAuth:
  - '[]'
securityDefinitions:
  apiKey:
    in: header
    name: authorization
    type: apiKey
swagger: "2.0"